refactor BaseHttpHandler
